leave mazama (julia error)

+++ deps/CustomOps/ScatterAddOps/ScatterAddOps.cu
@@ -1,6 +1,6 @@
 #include "cuda.h"
 
-__global__ void ScatterAddOps_forward_kernel(double *out, const double *ipt, const int64 *ii,
+__global__ void ScatterAddOps_forward_kernel(double *out, const double *ipt, const long long *ii,
     const double *update, int d, int n){
     int p =  blockIdx.x *blockDim.x + threadIdx.x;
     if (p < n){
@@ -8,7 +8,7 @@ __global__ void ScatterAddOps_forward_kernel(double *out, const double *ipt, con
     }
  }
  
-void Gpu_ScatterAddOps_forward(double *out, const double *ipt, const int64 *ii,
+void Gpu_ScatterAddOps_forward(double *out, const double *ipt, const long long *ii,
     const double *update, int d, int n){
     cudaMemcpy(out, ipt, sizeof(double) * d, cudaMemcpyDeviceToDevice);
     ScatterAddOps_forward_kernel<<< (n-1)/64 + 1, 64 >>>(out, ipt, ii, update, d, n);
@@ -17,7 +17,7 @@ void Gpu_ScatterAddOps_forward(double *out, const double *ipt, const int64 *ii,
 
 __global__ void ScatterAddOps_backward_kernel(double *grad_ipt, double *grad_update, 
     const double *grad_out,
-      const double *out, const double *ipt, const int64 *ii,
+      const double *out, const double *ipt, const long long *ii,
      const double *update, int d, int n){
     
     int p =  blockIdx.x *blockDim.x + threadIdx.x;
@@ -29,7 +29,7 @@ __global__ void ScatterAddOps_backward_kernel(double *grad_ipt, double *grad_upd
  void Gpu_ScatterAddOps_backward(
    double *grad_ipt, double *grad_update, 
    const double *grad_out,
-     const double *out, const double *ipt, const int64 *ii,
+     const double *out, const double *ipt, const long long *ii,
     const double *update, int d, int n){
     cudaMemcpy(grad_ipt, grad_out, sizeof(double)*d, cudaMemcpyDeviceToDevice);
     ScatterAddOps_backward_kernel<<< (n-1)/64, 64 >>>(grad_ipt, grad_update, grad_out, out, ipt, ii,
